101-print_listint_safe.c

+++ 0x13-more_singly_linked_lists/101-print_listint_safe.c
@@ -12,7 +12,7 @@
 
 size_t print_listint_safe(const listint_t *head)
 {
-	int i = 0, loop = 0;
+	size_t i = 0, loop = 0;
 	const listint_t *tor, *hare;
 
 	hare = head;
